LayoutTests/crypto/webkitSubtle/aes-cbc-invalid-length-expected.txt

CONSOLE MESSAGE: line 22: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test an unsupported AES key length.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw AES key from string literal...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 176
PASS key.usages is ['decrypt', 'encrypt']
Using the key to encrypt plaintext...
Failed, as expected. Note that the spec doesn't appear to clearly define which step should fail.
PASS successfullyParsed is true

TEST COMPLETE
